+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/HandledInternalNotifyScenario.swift
@@ -20,6 +20,7 @@ import Bugsnag
                 ["method":"yes()"]
             ]
             report.attachCustomStacktrace(frames, withType: "unreal")
+            return true
         }
     }
 }