+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/ModifyBreadcrumbInNotify.swift
@@ -17,6 +17,7 @@ class ModifyBreadcrumbInNotify: Scenario {
                     crumb.message = "Cache locked"
                 }
             })
+            return true
         }
     }